+++ src/main/java/theAlley/db/CacheDB.java
@@ -1,0 +1,21 @@
+package theAlley.db;
+
+import lombok.Data;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
+
+//item table
+@Entity
+@Data
+@Table(name = "cache")
+public class CacheDB {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer id;
+    private String mapName;
+    private String currentRoomStatus;
+}
