+++ src/main/java/theAlley/db/CacheRepo.java
@@ -1,7 +1,7 @@
-package the.alley.model;
+package theAlley.db;
 
 import org.springframework.data.repository.CrudRepository;
 
 public interface CacheRepo extends CrudRepository<CacheDB, Integer> {
-	CacheDB findById(int intValue);
+    CacheDB findById(int intValue);
 }
