+++ src/main/java/theAlley/db/ItemsDB.java
@@ -1,0 +1,24 @@
+package theAlley.db;
+
+import lombok.Data;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
+
+//item table
+@Entity
+@Data
+@Table(name = "item")
+public class ItemsDB {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer id;
+    private String name;
+    private String description;
+    private Integer attack;
+    private Integer defense;
+
+}
