+++ src/main/java/theAlley/db/MapDB.java
@@ -1,0 +1,24 @@
+package theAlley.db;
+
+import lombok.Data;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
+
+//map tables
+@Entity
+@Data
+@Table(name = "map")
+public class MapDB {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer id;
+    private String name;
+    private String description;
+    private Integer items;
+    private Integer npcs;
+    private Integer users;
+}
