+++ src/main/java/theAlley/db/MapRepo.java
@@ -1,7 +1,7 @@
-package the.alley.model;
+package theAlley.db;
 
 import org.springframework.data.repository.CrudRepository;
 
 public interface MapRepo extends CrudRepository<MapDB, Integer> {
-	 //List<MapDB> findByName(String name);
+    //List<MapDB> findByName(String name);
 }
