+++ src/main/java/theAlley/db/NpcRepo.java
@@ -1,0 +1,8 @@
+package theAlley.db;
+
+import org.springframework.data.repository.CrudRepository;
+
+public interface NpcRepo extends CrudRepository<NpcDB, Integer> {
+    //List<NpcDB> findByName(String name);
+    //NpcDB findByLocation(Integer location);
+}
