+++ src/main/java/theAlley/db/UserRepo.java
@@ -1,7 +1,7 @@
-package the.alley.model;
+package theAlley.db;
 
 import org.springframework.data.repository.CrudRepository;
 
 public interface UserRepo extends CrudRepository<UserDB, Integer> {
-	UserDB findByName(String name);
+    UserDB findByName(String name);
 }
