+++ src/main/java/theAlley/exception/NotFound.java
@@ -1,14 +1,17 @@
-package the.alley.exception;
+package theAlley.exception;
 
 public class NotFound extends RuntimeException {
     public NotFound() {
     }
+
     public NotFound(String message) {
         super(message);
     }
+
     public NotFound(String message, Throwable cause) {
         super(message, cause);
     }
+
     public NotFound(Throwable cause) {
         super(cause);
     }
