+++ src/main/java/theAlley/service/DarknessConstants.java
@@ -1,0 +1,14 @@
+package theAlley.service;
+
+public final class DarknessConstants {
+    public static final String map_0 = "a dingy bar with a single bartender.";
+    public static final String map_1 = "a dark street corner...";
+    public static final String map_2 = "a dark alley...";
+    public static final String npc_0 = "Surly Bartender";
+    public static final String npc_1 = "Vagrant";
+    public static final String npc_2 = "Unpaid Nightwatchman";
+    public static final String item_0 = "sterling silver";
+    public static final String item_1 = "gold doubloon";
+    public DarknessConstants() {
+    }
+}
