+++ src/main/resources/application.properties
@@ -1,4 +1,4 @@
-server.port = 1234
+server.port=1234
 # logging.level.org.apache.http=DEBUG
 logging.level.org.springframework=DEBUG
 # logging.level.com.sta=DEBUG
@@ -19,34 +19,27 @@ logging.hibernate=DEBUG
 # The password used for DEV
 # server.ssl.key-store-password=welcome1
 # server.ssl.key-store-provider=SUN
-
 # Thymeleaf
-spring.thymeleaf.cache = false
+spring.thymeleaf.cache=false
 spring.thymeleaf.suffix=.html
 spring.thymeleaf.mode=HTML
 spring.thymeleaf.encoding=UTF-8
-
 # H2 DB set up
 spring.datasource.url=jdbc:h2:mem:testdatabase
 spring.datasource.username=sa
 spring.datasource.password=password
 spring.datasource.driverClassName=org.h2.Driver
-
-
 # jpa
 spring.jpa.hibernate.ddl-auto=create-drop
 spring.jpa.show-sql=true
-
 ##H2 dialect sets
 # spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.H2Dialect
 # spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
-
 # maria DB set up
 #spring.datasource.url=jdbc:mariadb://localhost:2525
 #spring.datasource.username=root
 #spring.datasource.password=password
 #spring.datasource.driverClassName=org.mariadb.jdbc.Driver
-
 #mongo set up
 #spring.data.mongodb.authentication-database= # Authentication database name.
 # spring.data.mongodb.database=the_alley
@@ -69,7 +62,6 @@ spring.jpa.show-sql=true
 # stripped before adding them to the entity manager)
 # The SQL dialect makes Hibernate generate better SQL for the chosen database
 # spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MariaDB103Dialect
-
 #security
 #security.user.name=maestro
 #security.user.password=welcome1
