+++ src/main/resources/templates/template_1.html
@@ -1,43 +1,45 @@
 <!DOCTYPE HTML>
 <html xmlns:th="http://www.thymeleaf.org">
 <head>
-	<title>${description}</title>
-    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
+    <title>${description}</title>
+    <meta content="text/html; charset=UTF-8" http-equiv="Content-Type"/>
 </head>
 <body>
-	<div class="mainContainer">
-			<h1 id="desc" th:text='${description}' th:value='${description}'/><br />
-			<p><span id="name" th:text='${name}' th:value='${name}'/></span></p><br />
-			<span id="location" th:value='${location}' />
-			<p>The following fellows are present:</p>
-			<br />
-			<tr th:each="instance : ${npcs}">
-    			<td id="${instance.key}" th:value="${instance.value}" th:text="${instance.value}"></td>
-			</tr>
-			<p>Users in room:</p>
-			<tr th:each="instance : ${users}">
-				<td id="${instance.key}" th:value="${instance.value}" th:text="${instance.value}"></td>
-			</tr> 
-			<br />
-		<input id="textBox" type="text" width="50px" value="" /><br />
-		<input id="template1" type="button" onclick="variousInput();" width="15px" value="Various" />
-		<input id="template2" type="button" onclick="updateStatus();" width="15px" value="updateStatus" />
-	</div>
-	
-	
-	
-	<div id="output"></div><br />
-    <script type="text/javascript" th:src="@{js/jquery-3.6.0.js}" /></script>
-    <script type="text/javascript" th:src="@{js/client-side-script.js}" /></script>
-	<link rel="stylesheet" th:href="@{css/styles.css}"/>	
-	<script th:inline="javascript">
-	/*<![CDATA[*/
-	 	var mapName = /*[[${mapName}]]*/ 'User name';
-		var current_location = /*[[${location}]]*/ 'location';
-		var name = /*[[${mapName}]]*/ 'name';
-	 
-	/*]]>*/
-	</script>	
+<div class="mainContainer">
+    <h1 id="desc" th:text='${description}' th:value='${description}'/><br/>
+    <p><span id="name" th:text='${name}' th:value='${name}'/></span></p><br/>
+    <span id="location" th:value='${location}'/>
+    <p>The following fellows are present:</p>
+    <br/>
+    <tr th:each="instance : ${npcs}">
+        <td id="${instance.key}" th:text="${instance.value}" th:value="${instance.value}"></td>
+    </tr>
+    <p>Users in room:</p>
+    <tr th:each="instance : ${users}">
+        <td id="${instance.key}" th:text="${instance.value}" th:value="${instance.value}"></td>
+    </tr>
+    <br/>
+    <input id="textBox" type="text" value="" width="50px"/><br/>
+    <input id="template1" onclick="variousInput();" type="button" value="Various" width="15px"/>
+    <input id="template2" onclick="updateStatus();" type="button" value="updateStatus" width="15px"/>
+</div>
+
+
+<div id="output"></div>
+<br/>
+<script th:src="@{js/jquery-3.6.0.js}" type="text/javascript"/>
+</script>
+<script th:src="@{js/client-side-script.js}" type="text/javascript"/>
+</script>
+<link rel="stylesheet" th:href="@{css/styles.css}"/>
+<script th:inline="javascript">
+    /*<![CDATA[*/
+         var mapName = /*[[${mapName}]]*/ 'User name';
+        var current_location = /*[[${location}]]*/ 'location';
+        var name = /*[[${mapName}]]*/ 'name';
+
+    /*]]>*/
+</script>
 
 </body>
