+++ src/test/java/theAlley/the_alley/TheAlleyApplicationTests.java
@@ -1,12 +1,12 @@
-package the.alley.the_alley;
+package theAlley.the_alley;
 
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
 class TheAlleyApplicationTests {
 
-	//@Test
-	//void contextLoads() {
-	//}
+    //@Test
+    //void contextLoads() {
+    //}
 
 }
